pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bb2gf"
version = "1.0.0"
description = "Bitbucket Server/DC → GitFlic migrator"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32,<3",
  "python-dotenv>=1.0,<2",
  "PyYAML>=6,<7",
  "tenacity>=8.5,<9",
  "python-slugify>=8,<9",
  "click>=8.1.7,<9",
  "typer==0.12.3",
  "rich>=13.7,<14",
]

[project.scripts]
bb2gf = "src.main:app"

[tool.setuptools.packages.find]
include = ["src*"]
